requirements.txt
Flask
python-dateutil
poppler-utils
pytesseract
gunicorn
pillow
image
pdf2image
scikit-learn==1.0.1
pickle5
regex
nltk
spacy
werkzeug
